further changes

+++ migrations/20240920132747_products_table.sql
@@ -3,12 +3,14 @@
 CREATE TABLE IF NOT EXISTS products (
         id SERIAL PRIMARY KEY,
         name VARCHAR(50) NOT NULL,
-        price INT,
-        stock_count INT,
-    );
+        price NUMERIC(10, 2),
+        stock_count INT
+);
+
 -- +goose StatementEnd
 
 -- +goose Down
 -- +goose StatementBegin
 DROP TABLE products;
 -- +goose StatementEnd
+
